+++ cvxpy/constraints/__init__.py
@@ -22,6 +22,7 @@ from cvxpy.constraints.eq_constraint import EqConstraint
 from cvxpy.constraints.exponential import ExpCone
 from cvxpy.constraints.int_constr import IntConstr
 from cvxpy.constraints.leq_constraint import LeqConstraint
+from cvxpy.constraints.psd_constraint import PSDConstraint
 from cvxpy.constraints.second_order import SOC
 from cvxpy.constraints.semidefinite import SDP
 from cvxpy.constraints.soc_elemwise import SOC_Elemwise