+++ cvxpy/expressions/variables/__init__.py
@@ -21,3 +21,4 @@ from cvxpy.expressions.variables.bool_var import Bool
 from cvxpy.expressions.variables.int_var import Int
 from cvxpy.expressions.variables.variable import Variable
 from cvxpy.expressions.variables.semidef_var import Semidef
+from cvxpy.expressions.variables.symmetric import Symmetric