+++ cvxpy/tests/test_problem.py
@@ -21,7 +21,7 @@ from fractions import Fraction
 import cvxpy.settings as s
 from cvxpy.atoms import *
 from cvxpy.expressions.constants import Constant, Parameter
-from cvxpy.expressions.variables import Variable, Semidef, Bool
+from cvxpy.expressions.variables import Variable, Semidef, Bool, Symmetric
 from cvxpy.problems.objective import *
 from cvxpy.problems.problem import Problem
 from cvxpy.problems.solvers.utilities import SOLVERS, installed_solvers
@@ -1197,6 +1197,53 @@ class TestProblem(BaseTest):
         x0.value = 1
         prob.solve()
         self.assertAlmostEqual(prob.value, 1)
+
+    def test_psd_constraints(self):
+        """Test positive definite constraints.
+        """
+        # C = Variable(3, 3)
+        # obj = Maximize(C[0, 2])
+        # constraints = [diag(C) == 1,
+        #                C[0, 1] == 0.6,
+        #                C[1, 2] == -0.3,
+        #                C == C.T,
+        #                C >> 0]
+        # prob = Problem(obj, constraints)
+        # result = prob.solve()
+        # self.assertAlmostEqual(result, 0.583151)
+
+        # C = Variable(2, 2)
+        # obj = Maximize(C[0, 1])
+        # constraints = [C == 1, C >> [[2,0], [0, 2]]]
+        # prob = Problem(obj, constraints)
+        # result = prob.solve()
+        # self.assertEqual(prob.status, s.INFEASIBLE)
+
+        # C = Symmetric(2, 2)
+        # obj = Minimize(C[0, 0])
+        # constraints = [C << [[2,0], [0, 2]]]
+        # prob = Problem(obj, constraints)
+        # result = prob.solve(solver=s.CVXOPT)
+        # self.assertEqual(prob.status, s.UNBOUNDED)
+
+        C = Symmetric(2, 2)
+        obj = Maximize(C[0, 0])
+        constraints = [C << [[2,0], [0, 2]]]
+        prob = Problem(obj, constraints)
+        result = prob.solve(solver=s.SCS)
+        self.assertAlmostEqual(result, 2)
+
+        # Test the dual values.
+        psd_constr_dual = constraints[0].dual_value
+        C = Symmetric(2, 2)
+        X = Semidef(2)
+        obj = Maximize(C[0, 0])
+        constraints = [X == [[2,0], [0, 2]] - C]
+        prob = Problem(obj, constraints)
+        result = prob.solve(solver=s.SCS)
+        print psd_constr_dual
+        print constraints[0].dual_value
+        self.assertItemsAlmostEqual(constraints[0].dual_value, psd_constr_dual)
 
     def test_geo_mean(self):
         import numpy as np